requirements.txt
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
diffusers==0.32.2
numpy==1.26.4
transformers==4.50.3
accelerate==1.5.2
Pillow==10.4.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
opencv_python==4.10.0.84
timm==1.0.15
sentencepiece==0.2.0
ftfy
regex
timm
decord
einops
./sageattention-2.1.1-cp312-cp312-linux_x86_64.whl
